rename ring -> atomic
